# spring-ai-alibaba-graph/spring-ai-alibaba-graph-studio/src/main/java/com/alibaba/cloud/ai/api/DSLAPI.java
package com.alibaba.cloud.ai.api;

import com.alibaba.cloud.ai.common.R;
import com.alibaba.cloud.ai.exception.NotImplementedException;
import com.alibaba.cloud.ai.exception.SerializationException;
import com.alibaba.cloud.ai.model.App;
import com.alibaba.cloud.ai.param.DSLParam;
import com.alibaba.cloud.ai.saver.AppSaver;
import com.alibaba.cloud.ai.service.dsl.DSLAdapter;
import com.alibaba.cloud.ai.service.dsl.DSLDialectType;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.tags.Tag;
import org.springframework.core.io.ByteArrayResource;
import org.springframework.core.io.Resource;
import org.springframework.http.HttpHeaders;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;

import java.io.IOException;
import java.nio.charset.StandardCharsets;
import java.util.Optional;

@Tag(name = "DSL", description = "the DSL API")
public interface DSLAPI {

	DSLAdapter getAdapter(String dialect);

	AppSaver getAppSaver();

	@Operation(summary = "export app to dsl", tags = { "DSL" })
	@GetMapping(value = "/export/{id}", produces = "application/json")
	default R<String> exportDSL(@PathVariable("id") String id, @RequestParam("dialect") String dialect) {
		App app = Optional.ofNullable(getAppSaver().get(id))
			.orElseThrow(() -> new IllegalArgumentException("App not found: " + id));
		return R.success(getAdapter(dialect).exportDSL(app));
	}

	@Operation(summary = "export app to dsl file", tags = { "DSL" })
	@GetMapping(value = "/export-file/{id}", produces = MediaType.APPLICATION_OCTET_STREAM_VALUE)
	default ResponseEntity<Resource> exportDSLFile(@PathVariable("id") String id, @RequestParam("dialect") String dialect) {
		App app = Optional.ofNullable(getAppSaver().get(id))
				.orElseThrow(() -> new IllegalArgumentException("App not found: " + id));
		String dslContent = getAdapter(dialect).exportDSL(app);
		DSLDialectType dialectType = DSLDialectType.fromValue(dialect)
				.orElseThrow(()-> new NotImplementedException("Unsupported dsl dialect: " + dialect));
		ByteArrayResource resource = new ByteArrayResource(dslContent.getBytes(StandardCharsets.UTF_8));
		String fileName = app.getMetadata().getName() + dialectType.fileExtension();
		return ResponseEntity.ok()
				.header(HttpHeaders.CONTENT_DISPOSITION, "attachment; filename=\"" + fileName + "\"")
				.contentType(MediaType.APPLICATION_OCTET_STREAM)
				.contentLength(resource.contentLength())
				.body(resource);
	}

	@Operation(summary = "import app from dsl", tags = { "DSL" })
	@PostMapping(value = "/import", produces = "application/json")
	default R<App> importDSL(@RequestBody DSLParam param) {
		App app = getAdapter(param.getDialect()).importDSL(param.getContent());
		app = getAppSaver().save(app);
		return R.success(app);
	}

	@Operation(summary = "import app from dsl file ", tags = { "DSL" })
	@PostMapping(value = "/import-file", consumes = MediaType.MULTIPART_FORM_DATA_VALUE, produces = "application/json")
	default R<App> importDSLFile(
			@RequestPart("file") MultipartFile file,
			@RequestParam("dialect") String dialect) {
		String content;
		try {
			content = new String(file.getBytes(), StandardCharsets.UTF_8);

		}catch (IOException e){
			throw new SerializationException("Read dsl file failed, please check if the encoding of file is UTF_8 ");
		}

		App app = getAdapter(dialect).importDSL(content);
		app = getAppSaver().save(app);

		return R.success(app);
	}

}