# spring-ai-alibaba-graph/spring-ai-alibaba-graph-studio/src/main/java/com/alibaba/cloud/ai/model/App.java
package com.alibaba.cloud.ai.model;

import com.alibaba.cloud.ai.exception.NotImplementedException;
import com.alibaba.cloud.ai.service.run.RunnableModel;
import com.alibaba.cloud.ai.service.run.RunnableType;
import lombok.Data;

/**
 * Unified app model definition.
 */
@Data
public class App implements RunnableModel {

	private AppMetadata metadata;

	/**
	 * Spec has different implementations depending on the type of application. e.g.
	 * Workflow
	 */
	private Object spec;

	public App(AppMetadata metadata, Object spec) {
		this.metadata = metadata;
		this.spec = spec;
	}

	@Override
	public String id() {
		return metadata.getId();
	}

	@Override
	public RunnableType runnableType() {
		return RunnableType.fromValue(metadata.getMode())
				.orElseThrow(()->new NotImplementedException("Unsupported runnable type: " + metadata.getMode()));
	}

}